==> notebooks/00_exploration.ipynb <==
#load Data
DATA_PATH = '../data/raw/survey_results_public.csv'

#load important Libs
import pandas as pd 
import numpy as np
import matplotlib as plt
import seaborn as sb
pd.options.display.max_rows =10000

df = pd.read_csv(DATA_PATH)
df.head(5)

df.shape

# Display random answer
# observations: Multiple answers need to be splitted
# Reference to the schema needed to understand
df.sample(1).iloc[0]

df.describe()

# Investigate the questionable objects columns
questionable_cols = ['YearsCodePro', 'YearsCode', 'Age1stCode']

for col in questionable_cols: 
    print(col)
    print(df[col].unique().tolist())
    print('--------------------------')
    print()
