==> notebooks/05a_prediction_notebook.ipynb <==
LOG_DATA_PKL    =  "data.pkl"
LOG_MODEL_PKL   =  "model.pkl"

TRACKING_URI = "...\notebooks\mlruns"  # track file on lab
EXPERIMENT_ID = "1"
RUN_ID = "4c35a9c9d26a48e8bfccfae05a63d348" #of model

import os 
import sklearn
import pickle
import yaml

import pandas as pd

import mlflow
from mlflow.tracking import MlflowClient

## load model

artifact_path = os.path.join(TRACKING_URI.replace("file://", ""), 
                             EXPERIMENT_ID, 
                             RUN_ID, 
                             'artifacts')

# Load data pkl
data_path  = os.path.join(artifact_path, LOG_DATA_PKL)
with open(data_path, 'rb') as handle:
    data_pkl = pickle.load(handle)

# Load model pkl
model_path = os.path.join(artifact_path, LOG_MODEL_PKL)
with open(model_path, 'rb') as handle:
    model_pkl = pickle.load(handle)

model = model_pkl["model_object"]
